+++ continuous_integration/install-deps-platformio.sh
@@ -36,6 +36,9 @@ pio lib -g install adafruit/'Adafruit AM2315'
 echo "\e[32mInstalling adafruit/'Adafruit BME280 Library'\e[0m"
 pio lib -g install adafruit/'Adafruit BME280 Library'
 
+echo "\e[32mInstalling adafruit/'Adafruit BMP3XX Library'\e[0m"
+pio lib -g install adafruit/'Adafruit BMP3XX Library'
+
 echo "\e[32mInstalling adafruit/'DHT sensor library'\e[0m"
 pio lib -g install adafruit/'DHT sensor library'