+++ src/sensors/MeaSpecMS5803.h
@@ -43,7 +43,7 @@
  * the sensor can be stopped between measurements.
  *
  * @warning These I2C addresses are the same as those available for the Bosch
- * BME280 Barometric Pressure Sensor!  If you are also using one of those
+ * BME280, BMP388, and BMP390 sensors!  If you are also using one of those
  * sensors, make sure that the address for that sensor does not conflict with
  * the address of this sensor.
  *